One less crate
