+++ README.md
@@ -1,0 +1,4 @@
+# scoped mutex
+
+When a mutex and a closure love each other very much.
+
