+++ source/scoped-mutex/Cargo.toml
@@ -7,14 +7,14 @@ edition = "2021"
 version = "0.1"
 path = "../scoped-mutex-traits"
 
-[dependencies.scoped-mutex-impls]
-version = "0.1"
-path = "../scoped-mutex-impls"
+[dependencies.critical-section]
+version = "1.1"
+optional = true
 
 [features]
 default = [
     "impl-critical-section",
 ]
-impl-critical-section = ["scoped-mutex-impls/impl-critical-section"]
-impl-unsafe-cortex-m-single-core = ["scoped-mutex-impls/impl-unsafe-cortex-m-single-core"]
+impl-critical-section = ["dep:critical-section"]
+impl-unsafe-cortex-m-single-core = []
 std = []
